% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HLAalignments}
\alias{HLAalignments}
\title{Amino acid, cDNA, or gDNA sequence alignments for the genes in the ANHIG/IMGTHLA GitHub repository.}
\format{
A large list of 4 sublists that include dataframes of amino-acid, cDNA and gDNA aligmnents.
\itemize{
\item(prot: data frames of amino-acid sequence alignments)
\item(nuc: data frames of nucleotide sequence alignments)
\item(gen: data frames of genomic sequence alignments)
\item(version: The IPD-IMGT/HLA Database release version under which these data were generated)
}
}
\source{
https://github.com/ANHIG/IMGTHLA/tree/Latest/alignments
}
\usage{
data(HLAalignments)
}
\description{
A list object of sub-lists of R dataframes of amino-acid, cDNA and gDNA sequence alignments for all of the HLA and HLA-region genes supported in the ANHIG/IMGTHLA GitHub Repository.
This object is built by the alignmentFull() function in the HLAtools package.
}
\keyword{datasets}
